Command:Module, with commands defined separately
Instead of Command being a class, it is now a module. Commands are
defined individually, each adding a method to the module, in
/lib/hs/commands.
